backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
gunicorn<=20.1.0
websockets==10.0
numpy
pyyaml
imageio
aiofiles
python-multipart
Cython
future
scikit-image
sklearn
pynwb
ipyparallel
holoviews
psutil
nbstripout
tensorflow<=2.6.1
torch<=1.10.0
opencv-python
suite2p
statsmodels<=0.13.1
snakemake